+++ Resources/Prototypes/_RMC14/Entities/StatusEffects/rmc_hushed.yml
@@ -1,8 +1,8 @@
 # Forces entity to only whisper instead of speaking normally (RMC14)
 - type: entity
   parent: MobStatusEffectBase
-  id: RMCStatusEffectWhisperOnly
-  name: whisper only
+  id: RMCStatusEffectHushed
+  name: hushed
   components:
   - type: StatusEffect
-    alert: RMCWhisperOnly
+    alert: RMCHushed
